+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.8.5
 *Release date: 2015-03-??*
 
   * Add Button page element
+  * AppiumTestCase has a new attribute app_strings, a dict with application strings in the active language
 
 v0.8.4
 ------